Refactor FIleBrowser into new class in engine's editor
